+++ library/core/src/iter/adapters/fuse.rs
@@ -1,10 +1,11 @@
-use super::InPlaceIterable;
-use crate::intrinsics;
-use crate::iter::adapters::zip::try_get_unchecked;
-use crate::iter::adapters::SourceIter;
-use crate::iter::TrustedRandomAccess;
-use crate::iter::{DoubleEndedIterator, ExactSizeIterator, FusedIterator, Iterator};
-use crate::ops::Try;
+use crate::{
+    intrinsics,
+    iter::{
+        adapters::{zip::try_get_unchecked, InPlaceIterable, SourceIter},
+        DoubleEndedIterator, ExactSizeIterator, FusedIterator, Iterator, TrustedRandomAccess,
+    },
+    ops::Try,
+};
 
 /// An iterator that yields `None` forever after the underlying iterator
 /// yields `None` once.
