+++ library/core/src/iter/adapters/zip.rs
@@ -1,9 +1,10 @@
-use crate::cmp;
-use crate::fmt::{self, Debug};
-
-use super::super::{
-    DoubleEndedIterator, ExactSizeIterator, FusedIterator, InPlaceIterable, Iterator, SourceIter,
-    TrustedLen,
+use crate::{
+    cmp,
+    fmt::{self, Debug},
+    iter::{
+        DoubleEndedIterator, ExactSizeIterator, FusedIterator, InPlaceIterable, Iterator,
+        SourceIter, TrustedLen,
+    },
 };
 
 /// An iterator that iterates two other iterators simultaneously.
@@ -21,7 +22,7 @@ pub struct Zip<A, B> {
     len: usize,
 }
 impl<A: Iterator, B: Iterator> Zip<A, B> {
-    pub(in super::super) fn new(a: A, b: B) -> Zip<A, B> {
+    pub(in crate::iter) fn new(a: A, b: B) -> Zip<A, B> {
         ZipImpl::new(a, b)
     }
     fn super_nth(&mut self, mut n: usize) -> Option<(A::Item, B::Item)> {
